pyproject.toml
[tool.poetry]
name = "app"
version = "0.1.0"
description = ""
authors = ["LE THI HANG <lethihang111001@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.110.0"
pydantic = "^2.6.4"
uvicorn = "^0.29.0"
sqlalchemy = "^2.0.28"
psycopg2 = "^2.9.9"
werkzeug = "^3.0.1"
sqlalchemy-utils = "^0.41.1"
fastapi-jwt-auth = "^0.5.0"
jwt = "^1.3.1"
crypto = "^1.4.1"
python-jose = {version = "^3.3.0", extras = ["cryptography"]}
alembic = "^1.13.1"
bcrypt = "^4.1.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
